requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.35.13
deepgram-sdk>=3.0.0
numpy==1.26.4
scipy==1.13.1
soundfile==0.12.1
pyaudio==0.2.14
python-multipart==0.0.9
